can update names for context and timeCategory

+++ handlers/contextPage.py
@@ -6,32 +6,42 @@ from . import accessControl
 
 
 class ContextPage(Handler):
+    finished_events = {}
+    unfinished_events = {}
 
     @accessControl.user_logged_in
     @accessControl.context_exist
     def get(self, context_id, context):
-        finished_events = {}
-        unfinished_events = {}
+
         for event in context.events:
             if event.finished:
-                if not finished_events.get(str(event.time_exe_start.date())):
-                    finished_events[str(event.time_exe_start.date())] = [event]
+                if not self.finished_events.get(str(event.time_exe_start.date())):
+                    self.finished_events[str(event.time_exe_start.date())] = [event]
                 else:
-                    finished_events[str(event.time_exe_start.date())].append(event)
+                    self.finished_events[str(event.time_exe_start.date())].append(event)
             else:
-                if not unfinished_events.get(str(event.time_exe_start.date())):
-                    unfinished_events[str(event.time_exe_start.date())] = [event]
+                if not self.unfinished_events.get(str(event.time_exe_start.date())):
+                    self.unfinished_events[str(event.time_exe_start.date())] = [event]
                 else:
-                    unfinished_events[str(event.time_exe_start.date())].append(event)
+                    self.unfinished_events[str(event.time_exe_start.date())].append(event)
         self.render("contextPage.html",
             context_name=context.name,
-            finished_events=finished_events,
-            unfinished_events=unfinished_events)
+            finished_events=self.finished_events,
+            unfinished_events=self.unfinished_events)
 
     @accessControl.user_logged_in
     @accessControl.context_exist
     def post(self, context_id, context):
-        for event in context.events:
-            event.delete()
-        context.delete()
-        self.redirect("/contexts")
+        if 'Delete Context' in self.request.params:
+            for event in context.events:
+                event.delete()
+            context.delete()
+            self.redirect("/contexts")
+        else: #Update
+            context_name = self.request.get('context_name')
+            context.name = context_name
+            context.put()
+            self.render("contextPage.html",
+                context_name=context.name,
+                finished_events=self.finished_events,
+                unfinished_events=self.unfinished_events)
