+++ handlers/timeCategoryPage.py
@@ -6,32 +6,42 @@ from . import accessControl
 
 
 class TimeCategoryPage(Handler):
+    finished_events = {}
+    unfinished_events = {}
 
     @accessControl.user_logged_in
     @accessControl.timeCategory_exist
     def get(self, timeCategory_id, timeCategory):
-        finished_events = {}
-        unfinished_events = {}
+
         for event in timeCategory.events:
             if event.finished:
-                if not finished_events.get(str(event.time_exe_start.date())):
-                    finished_events[str(event.time_exe_start.date())] = [event]
+                if not self.finished_events.get(str(event.time_exe_start.date())):
+                    self.finished_events[str(event.time_exe_start.date())] = [event]
                 else:
-                    finished_events[str(event.time_exe_start.date())].append(event)
+                    self.finished_events[str(event.time_exe_start.date())].append(event)
             else:
-                if not unfinished_events.get(str(event.time_exe_start.date())):
-                    unfinished_events[str(event.time_exe_start.date())] = [event]
+                if not self.unfinished_events.get(str(event.time_exe_start.date())):
+                    self.unfinished_events[str(event.time_exe_start.date())] = [event]
                 else:
-                    unfinished_events[str(event.time_exe_start.date())].append(event)
+                    self.unfinished_events[str(event.time_exe_start.date())].append(event)
         self.render("timeCategoryPage.html",
             timeCategory_name=timeCategory.name,
-            finished_events=finished_events,
-            unfinished_events=unfinished_events)
+            finished_events=self.finished_events,
+            unfinished_events=self.unfinished_events)
 
     @accessControl.user_logged_in
     @accessControl.timeCategory_exist
     def post(self, timeCategory_id, timeCategory):
-        for event in timeCategory.events:
-            event.delete()
-        timeCategory.delete()
-        self.redirect("/timeCategories")
+        if 'Delete Category' in self.request.params:
+            for event in timeCategory.events:
+                event.delete()
+            timeCategory.delete()
+            self.redirect("/timeCategories")
+        else: #Update
+            category_name = self.request.get('timeCategory_name')
+            timeCategory.name = timeCategory_name
+            timeCategory.put()
+            self.render("timeCategoryPage.html",
+                timeCategory_name=timeCategory.name,
+                finished_events=self.finished_events,
+                unfinished_events=self.unfinished_events)
